Update midnight.js
When using anything but text in the button code, it is net clickable. For example, adding an icon (like FontAwesome) and clicking the icon, the click is not registered.

In the event listener, you're checking if the event target (the button) is the same as the darkModeButton. When the button only includes text, clicking on the text will still register as clicking the button.

The browser interprets the click event as targeting the icon, not the button. The event.target would refer to the icon (<i>), not the button (<button>), which fails the comparison event.target === darkModeButton.